requirements.txt
eventkit==1.0.3
ib-insync==0.9.86
kafka-python==2.0.2
nest-asyncio==1.6.0
numpy==1.26.3
PyYAML~=6.0.1
